Seperating out some code
